Move permissions module to separate file
